requirements.txt
cycler==0.10.0
Cython==0.29.21
kiwisolver==1.2.0
matplotlib==3.2.2
MouseInfo==0.1.3
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
python-dateutil==2.8.1
PyTweening==1.0.3
six==1.15.0
